Move stats definition to own file
